// laravel/resources/js/model/domain/raceModel.ts
export interface RaceCreation {
    name: string;
    manager: string;
    isCompetitive: boolean;
    duration: string;
    difficulty: string;
    minorPrice: number;
    majorPrice: number;
    minParticipants: number;
    illustration: string;
    startDate: string;
    startTime: string; // HH:mm
    endDate: string;
    endTime: string; // HH:mm
}